current back to int

+++ src/boot/ksdk1.1.0/devINA219.c
@@ -158,8 +158,8 @@ printSensorDataINA219(bool hexModeFlag, uint8_t address)
 	uint16_t	readSensorRegisterValueLSB;
 	uint16_t	readSensorRegisterValueMSB;
 	int16_t		readSensorRegisterValueCombined;
-	int16_t	  Voltage;
-	float 		Current;
+	int16_t		Voltage;
+	int16_t		Current;
 	int16_t 	Power;
 	WarpStatus	i2cReadStatus;
 
@@ -205,8 +205,8 @@ printSensorDataINA219(bool hexModeFlag, uint8_t address)
 		}
 		else if (address == 0x04)
 		{
-			Current = (float)readSensorRegisterValueCombined / 10.0f;
-			warpPrint("Current : %.2f mA,\n", Current);
+			Current = readSensorRegisterValueCombined / 10;
+			warpPrint("Current : %d mA,\n", Current);
 		}
 		else
 		{
